stabilize PolicyDefinitionAPI
