+++ extensions/control-plane/api/management-api/policy-definition-api/src/main/java/org/eclipse/edc/connector/controlplane/api/management/policy/v3/PolicyDefinitionApiV3Controller.java
@@ -25,18 +25,31 @@ import jakarta.ws.rs.Path;
 import jakarta.ws.rs.PathParam;
 import jakarta.ws.rs.Produces;
 import org.eclipse.edc.connector.controlplane.api.management.policy.BasePolicyDefinitionApiController;
+import org.eclipse.edc.connector.controlplane.api.management.policy.model.PolicyEvaluationPlanRequest;
+import org.eclipse.edc.connector.controlplane.api.management.policy.model.PolicyValidationResult;
+import org.eclipse.edc.connector.controlplane.policy.spi.PolicyDefinition;
 import org.eclipse.edc.connector.controlplane.services.spi.policydefinition.PolicyDefinitionService;
+import org.eclipse.edc.spi.EdcException;
 import org.eclipse.edc.spi.monitor.Monitor;
 import org.eclipse.edc.transform.spi.TypeTransformerRegistry;
 import org.eclipse.edc.validator.spi.JsonObjectValidatorRegistry;
+import org.eclipse.edc.web.spi.exception.InvalidRequestException;
+import org.eclipse.edc.web.spi.exception.ObjectNotFoundException;
+import org.eclipse.edc.web.spi.exception.ValidationFailureException;
+
+import java.util.ArrayList;
 
 import static jakarta.ws.rs.core.MediaType.APPLICATION_JSON;
+import static org.eclipse.edc.connector.controlplane.api.management.policy.model.PolicyEvaluationPlanRequest.EDC_POLICY_EVALUATION_PLAN_REQUEST_TYPE;
+import static org.eclipse.edc.web.spi.exception.ServiceResultHandler.exceptionMapper;
 
 @Consumes(APPLICATION_JSON)
 @Produces(APPLICATION_JSON)
 @Path("/v3/policydefinitions")
 public class PolicyDefinitionApiV3Controller extends BasePolicyDefinitionApiController implements PolicyDefinitionApiV3 {
-    public PolicyDefinitionApiV3Controller(Monitor monitor, TypeTransformerRegistry transformerRegistry, PolicyDefinitionService service, JsonObjectValidatorRegistry validatorRegistry) {
+
+    public PolicyDefinitionApiV3Controller(Monitor monitor, TypeTransformerRegistry transformerRegistry, PolicyDefinitionService service,
+                                           JsonObjectValidatorRegistry validatorRegistry) {
         super(monitor, transformerRegistry, service, validatorRegistry);
     }
 
@@ -72,5 +85,47 @@ public class PolicyDefinitionApiV3Controller extends BasePolicyDefinitionApiCont
     @Override
     public void updatePolicyDefinitionV3(@PathParam("id") String id, JsonObject input) {
         updatePolicyDefinition(id, input);
+    }
+
+    @POST
+    @Path("{id}/validate")
+    @Override
+    public JsonObject validatePolicyDefinitionV3(@PathParam("id") String id) {
+        var definition = service.findById(id);
+        if (definition == null) {
+            throw new ObjectNotFoundException(PolicyDefinition.class, id);
+        }
+
+        var messages = new ArrayList<String>();
+
+        var result = service.validate(definition.getPolicy())
+                .onFailure(failure -> messages.addAll(failure.getMessages()));
+
+        var validationResult = new PolicyValidationResult(result.succeeded(), messages);
+
+        return transformerRegistry.transform(validationResult, JsonObject.class)
+                .orElseThrow(f -> new EdcException("Error creating response body: " + f.getFailureDetail()));
+    }
+
+    @POST
+    @Path("{id}/evaluationplan")
+    @Override
+    public JsonObject createExecutionPlaneV3(@PathParam("id") String id, JsonObject request) {
+
+        validatorRegistry.validate(EDC_POLICY_EVALUATION_PLAN_REQUEST_TYPE, request).orElseThrow(ValidationFailureException::new);
+
+        var planeRequest = transformerRegistry.transform(request, PolicyEvaluationPlanRequest.class)
+                .orElseThrow(InvalidRequestException::new);
+
+        var definition = service.findById(id);
+        if (definition == null) {
+            throw new ObjectNotFoundException(PolicyDefinition.class, id);
+        }
+
+        var plan = service.createEvaluationPlan(planeRequest.policyScope(), definition.getPolicy())
+                .orElseThrow(exceptionMapper(PolicyDefinition.class, definition.getId()));
+
+        return transformerRegistry.transform(plan, JsonObject.class)
+                .orElseThrow(f -> new EdcException("Error creating response body: " + f.getFailureDetail()));
     }
 }
